Added more functionality
Added Menus as well as embedded Alpha Vantage Stock API which will make getting historical stock data a lot easier than scraping yahoo finance. Alpha Vantage has a-lot of pros and helpful data as compared to Yahoo finance, which was  a pain just to scrape a title of the stock and a current price. This commit is only a partial of the working menu. There is 6 options i have for the main menu:
1. Option Payouts
2. News - which could be an overall Newspaper type of feed, or a feed pertaining to a certain stock
3. Watchlist Creation - a function to create a watchlist for the day going on forward
4. Stock Info  - would show information when its aftermarket hours
5. Live Stock Info - would show information for active market hours, refresh counters for every min, 5mins 15 mins, 30 mins, 60 mins.
6. Settings - which i would implement later on so users can input their own keys for example, and other miscellaneous settings